fix: prevent negative the number of request

+++ limiter.go
@@ -6,7 +6,10 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-func LimitMiddleware(limit int) gin.HandlerFunc {
+// LimitMiddleware is a middleware that limits the number of requests at the same time.
+//
+// If the number of requests exceeds the limit, the middleware will return a 429 Too Many Requests status.
+func LimitMiddleware(limit uint64) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		ch := make(chan struct{}, limit)
